moved interfaces to own file
